LRCI-1791 Extract inner classes and create Payload class hierarchy
