+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/github/webhook/GitHubWebhookPayloadProcessor.java
@@ -2407,351 +2407,10 @@ public class GitHubWebhookPayloadProcessor {
 	private List<String> _jiraProjectKeys;
 	private Pattern _passingTestSuiteStatusDescriptionPattern = Pattern.compile(
 		"\"ci:test:(?<testSuiteName>[^\"]+)\"\\s*has PASSED.");
-	private final PayloadJSONObject _payloadJSONObject;
+	private final Payload _payload;
 	private final Map<String, Long> _testPullRequestQueryStrings =
 		new ConcurrentHashMap<>(100);
 	private final Map<String, Long> _testPullRequestURLs =
 		new ConcurrentHashMap<>(100);
-
-	private static class PayloadJSONObject extends JSONObject {
-
-		public PayloadJSONObject(String source) {
-			super(_cleanupSource(source));
-		}
-
-		public PullRequest getPullRequest() {
-			if (_pullRequest == null) {
-				String path = "issue/pull/pull_request/html_url";
-
-				if (!has("issue")) {
-					path = "pull_request/_links/html/href";
-				}
-
-				_pullRequest = new PullRequest(getStringByPath(path));
-			}
-
-			return _pullRequest;
-		}
-
-		public String getPusherBranch() {
-			if (!isPusher()) {
-				throw new IllegalStateException(
-					"Unable to find branch in a non-pusher payload");
-			}
-
-			Matcher matcher = _branchPattern.matcher(getString("ref"));
-
-			if (matcher.matches()) {
-				return matcher.group(1);
-			}
-
-			return null;
-		}
-
-		public String getStringByPath(String path) {
-			return _getStringByPath(this, path);
-		}
-
-		public boolean isCommentCreated() {
-			String action = getAction();
-
-			if (!action.equals("created")) {
-				return false;
-			}
-
-			if (optJSONObject("comment") == null) {
-				return false;
-			}
-
-			return true;
-		}
-
-		public boolean isPullRequestOpened() {
-			String pullRequestAction = getPullRequestAction();
-
-			if (pullRequestAction.equals("opened")) {
-				return true;
-			}
-
-			return false;
-		}
-
-		public boolean isPullRequestSynchronize() {
-			String pullRequestAction = getPullRequestAction();
-
-			if (pullRequestAction.equals("synchronize")) {
-				return true;
-			}
-
-			return false;
-		}
-
-		public boolean isPusher() {
-			if (optJSONObject("pusher") != null) {
-				return true;
-			}
-
-			return false;
-		}
-
-		protected String getAction() {
-			String action = optString("action");
-
-			if (action == null) {
-				action = "";
-			}
-
-			return action;
-		}
-
-		protected String getPullRequestAction() {
-			if (optJSONObject("pull_request") == null) {
-				return "";
-			}
-
-			return getAction();
-		}
-
-		private static String _cleanupSource(String source) {
-			StringBuilder sb = new StringBuilder();
-
-			while (!source.isEmpty()) {
-				int start = source.indexOf("\"");
-
-				if (start == -1) {
-					sb.append(source);
-
-					source = "";
-
-					continue;
-				}
-
-				sb.append(source.substring(0, start));
-
-				int end = source.indexOf("\"", start + 1);
-
-				if (end == -1) {
-					throw new IllegalArgumentException(
-						"Unterminated quote found after index " + start);
-				}
-
-				String quotedString = source.substring(start, end + 1);
-
-				if (quotedString.contains("\n")) {
-					quotedString = quotedString.replaceAll("\n", "");
-
-					System.out.println("quotedString:\n" + quotedString);
-				}
-
-				sb.append(quotedString);
-
-				source = source.substring(end + 1);
-			}
-
-			return sb.toString();
-		}
-
-		private String _getStringByPath(JSONObject jsonObject, String path) {
-			Matcher matcher = _jsonPathPattern.matcher(path);
-
-			matcher.find();
-
-			String key = matcher.group(1);
-
-			String remainingPath = matcher.group(2);
-
-			if ((remainingPath == null) || remainingPath.isEmpty()) {
-				return jsonObject.getString(key);
-			}
-
-			return _getStringByPath(
-				jsonObject.getJSONObject(key), remainingPath);
-		}
-
-		private static final Pattern _branchPattern = Pattern.compile(
-			"refs/heads/(.*)");
-		private static final Pattern _jsonPathPattern = Pattern.compile(
-			"([^/]+)/*(.*)");
-
-		private PullRequest _pullRequest;
-
-	}
-
-	private static class PullRequestTesterParameters
-		extends HashMap<String, String> {
-
-		public PullRequestTesterParameters(PullRequest pullRequest) {
-			_pullRequest = pullRequest;
-
-			Properties jenkinsBuildProperties;
-
-			try {
-				jenkinsBuildProperties =
-					JenkinsResultsParserUtil.getBuildProperties();
-			}
-			catch (IOException ioException) {
-				throw new RuntimeException(
-					"Unable to get build properties", ioException);
-			}
-
-			put(_KEY_CI_TEST_SUITE_NAME, "default");
-
-			put(
-				_KEY_JENKINS_AUTHENTICATION_TOKEN,
-				jenkinsBuildProperties.getProperty(
-					"jenkins.authentication.token"));
-
-			String ciForwardReceiverUsername =
-				jenkinsBuildProperties.getProperty(
-					"pull.request.forward.default.receiver.username");
-
-			if (!JenkinsResultsParserUtil.isNullOrEmpty(
-					ciForwardReceiverUsername)) {
-
-				setCiForwardReceiverUsername(ciForwardReceiverUsername);
-			}
-		}
-
-		public String getCiForwardReceiverUsername() {
-			return get(_KEY_CI_FORWARD_RECEIVER_USERNAME);
-		}
-
-		public String getCiReevaluateBuildId() {
-			return get(_KEY_CI_REEVALUATE_BUILD_ID);
-		}
-
-		public String getCiTestSuiteName() {
-			return get(_KEY_CI_TEST_SUITE_NAME);
-		}
-
-		public PullRequest getPullRequest() {
-			return _pullRequest;
-		}
-
-		@Override
-		public String put(String key, String value) {
-			if (JenkinsResultsParserUtil.isNullOrEmpty(value) &&
-				containsKey(key)) {
-
-				return remove(key);
-			}
-
-			return super.put(key, value);
-		}
-
-		public void setCiForwardReceiverUsername(
-			String ciForwardReceiverUsername) {
-
-			put(_KEY_CI_FORWARD_RECEIVER_USERNAME, ciForwardReceiverUsername);
-		}
-
-		public void setCiReevaluateBuildId(String ciReevaluateBuildId) {
-			put(_KEY_CI_REEVALUATE_BUILD_ID, ciReevaluateBuildId);
-		}
-
-		public void setCiTestSuiteName(String ciTestSuiteName) {
-			if (JenkinsResultsParserUtil.isNullOrEmpty(ciTestSuiteName)) {
-				ciTestSuiteName = "default";
-			}
-
-			put(_KEY_CI_TEST_SUITE_NAME, ciTestSuiteName);
-		}
-
-		public void setPortalBundlesDistURL(String portalBundlesDistURL) {
-			put(_KEY_PORTAL_BUNDLES_DIST_URL, portalBundlesDistURL);
-		}
-
-		public void setUpstreamBranchSHA(String upstreamBranchSHA) {
-			put(_KEY_GITHUB_UPSTREAM_BRANCH_SHA, upstreamBranchSHA);
-		}
-
-		public String toQueryString() {
-			StringBuilder sb = new StringBuilder();
-
-			for (Map.Entry<String, String> entry : entrySet()) {
-				_appendParameter(sb, entry.getKey(), entry.getValue());
-			}
-
-			_appendParameter(
-				sb, _KEY_PULL_REQUEST_NUMBER, _pullRequest.getNumber());
-			_appendParameter(
-				sb, "GITHUB_ORIGIN_NAME",
-				_encode(_pullRequest.getSenderUsername()));
-			_appendParameter(
-				sb, "GITHUB_RECEIVER_USERNAME",
-				_pullRequest.getReceiverUsername());
-			_appendParameter(
-				sb, "GITHUB_SENDER_BRANCH_NAME",
-				_pullRequest.getSenderBranchName());
-			_appendParameter(
-				sb, "GITHUB_SENDER_BRANCH_SHA", _pullRequest.getSenderSHA());
-			_appendParameter(
-				sb, "GITHUB_SENDER_USERNAME",
-				_encode(_pullRequest.getSenderUsername()));
-			_appendParameter(
-				sb, "GITHUB_UPSTREAM_BRANCH_NAME",
-				_pullRequest.getUpstreamRemoteGitBranchName());
-
-			if (!containsKey(_KEY_GITHUB_UPSTREAM_BRANCH_SHA)) {
-				_appendParameter(
-					sb, _KEY_GITHUB_UPSTREAM_BRANCH_SHA,
-					_pullRequest.getUpstreamBranchSHA());
-			}
-
-			_appendParameter(sb, "PULL_REQUEST_URL", _pullRequest.getHtmlURL());
-			_appendParameter(
-				sb, "REPOSITORY_NAME", _pullRequest.getGitRepositoryName());
-
-			return sb.toString();
-		}
-
-		private void _appendParameter(
-			StringBuilder sb, String name, String value) {
-
-			if (JenkinsResultsParserUtil.isNullOrEmpty(value)) {
-				return;
-			}
-
-			if (sb.length() > 0) {
-				sb.append("&");
-			}
-
-			sb.append(name);
-			sb.append("=");
-			sb.append(value);
-		}
-
-		private String _encode(String string) {
-			try {
-				return URLEncoder.encode(string, "UTF-8");
-			}
-			catch (UnsupportedEncodingException unsupportedEncodingException) {
-				throw new RuntimeException(
-					"Unsupported Encoding: UTF-8",
-					unsupportedEncodingException);
-			}
-		}
-
-		private static final String _KEY_CI_FORWARD_RECEIVER_USERNAME =
-			"CI_FORWARD_RECEIVER_USERNAME";
-
-		private static final String _KEY_CI_REEVALUATE_BUILD_ID = "BUILD_ID";
-
-		private static final String _KEY_CI_TEST_SUITE_NAME = "CI_TEST_SUITE";
-
-		private static final String _KEY_GITHUB_UPSTREAM_BRANCH_SHA =
-			"GITHUB_UPSTREAM_BRANCH_SHA";
-
-		private static final String _KEY_JENKINS_AUTHENTICATION_TOKEN = "token";
-
-		private static final String _KEY_PORTAL_BUNDLES_DIST_URL =
-			"PORTAL_BUNDLES_DIST_URL";
-
-		private static final String _KEY_PULL_REQUEST_NUMBER =
-			"GITHUB_PULL_REQUEST_NUMBER";
-
-		private PullRequest _pullRequest;
-
-	}
 
 }
