+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/github/webhook/PullRequestCommentPayload.java
@@ -1,0 +1,48 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.jenkins.results.parser.github.webhook;
+
+import com.liferay.jenkins.results.parser.PullRequest;
+
+import org.json.JSONObject;
+
+/**
+ * @author Peter Yoo
+ */
+public class PullRequestCommentPayload extends PullRequestPayload {
+
+	public PullRequestCommentPayload(JSONObject jsonObject) {
+		super(jsonObject);
+
+		comment = new PullRequest.Comment(jsonObject.getJSONObject("comment"));
+	}
+
+	public PullRequest.Comment getComment() {
+		return comment;
+	}
+
+	@Override
+	public PullRequest getPullRequest() {
+		if (pullRequest == null) {
+			pullRequest = new PullRequest(
+				get("issue/pull/pull_request/html_url"));
+		}
+
+		return pullRequest;
+	}
+
+	protected PullRequest.Comment comment;
+
+}
