+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/github/webhook/PullRequestPayload.java
@@ -1,0 +1,40 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.jenkins.results.parser.github.webhook;
+
+import com.liferay.jenkins.results.parser.PullRequest;
+
+import org.json.JSONObject;
+
+/**
+ * @author Peter Yoo
+ */
+public class PullRequestPayload extends Payload {
+
+	public PullRequestPayload(JSONObject jsonObject) {
+		super(jsonObject);
+	}
+
+	public PullRequest getPullRequest() {
+		if (pullRequest == null) {
+			pullRequest = new PullRequest(get("pull_request/_links/html/href"));
+		}
+
+		return pullRequest;
+	}
+
+	protected PullRequest pullRequest;
+
+}
